requirement.txt
numpy>=1.24.0
requests>=2.28.0
PySide6>=6.5.0
PyYAML>=6.0
pyaudio>=0.2.11
scipy>=1.10.0
soundfile>=0.12.0
onnxruntime>=1.15.0
sounddevice>=0.5.2
spotipy>=2.25.1
ruamel.yaml>=0.17.0
openai>=1.3.0